Removed unneeded DevelopmentModeBaseCommands class.
